refactor: split actions
