+++ pokedex/src/store/index.js
@@ -8,5 +8,3 @@ export const store = createStore(
   rootReducer,
   composeWithDevTools(applyMiddleware(thunk)),
 );
-
-// TODO: Переименовать скрипт
